Fix Automate tests for 5.8
