+++ cfme/automate/__init__.py
@@ -54,27 +54,3 @@ class AutomateCustomization(CFMENavigateStep):
 
     def step(self):
         self.view.navigation.select(*automate_menu_name(self.obj.appliance) + ['Customization'])
-
-
-class AutomateExplorerView(BaseLoggedInPage):
-    @property
-    def is_displayed(self):
-        return (
-            self.logged_in_as_current_user and
-            self.navigation.currently_selected == automate_menu_name(
-                self.context['object'].appliance) + ['Explorer'])
-
-    @View.nested
-    class datastore(Accordion):  # noqa
-        tree = ManageIQTree()
-
-    configuration = Dropdown('Configuration')
-
-
-@navigator.register(Server)
-class AutomateExplorer(CFMENavigateStep):
-    VIEW = AutomateExplorerView
-    prerequisite = NavigateToSibling('LoggedIn')
-
-    def step(self):
-        self.view.navigation.select(*automate_menu_name(self.obj.appliance) + ['Explorer'])
