+++ cfme/automate/explorer/__init__.py
@@ -1,11 +1,15 @@
 # -*- coding: utf-8 -*-
 import re
 
+from navmazing import NavigateToSibling
 from widgetastic.widget import View
 from widgetastic_manageiq import Accordion, ManageIQTree
 from widgetastic_patternfly import Dropdown, FlashMessages
 
 from cfme import BaseLoggedInPage
+from cfme.base import Server
+from cfme.base.ui import automate_menu_name
+from utils.appliance.implementations.ui import navigator, CFMENavigateStep
 
 
 class AutomateExplorerView(BaseLoggedInPage):
@@ -15,7 +19,8 @@ class AutomateExplorerView(BaseLoggedInPage):
     def in_explorer(self):
         return (
             self.logged_in_as_current_user and
-            self.navigation.currently_selected == ['Automate', 'Explorer'])
+            self.navigation.currently_selected == automate_menu_name(
+                self.context['object'].appliance) + ['Explorer'])
 
     @property
     def is_displayed(self):
@@ -27,6 +32,15 @@ class AutomateExplorerView(BaseLoggedInPage):
         tree = ManageIQTree()
 
     configuration = Dropdown('Configuration')
+
+
+@navigator.register(Server)
+class AutomateExplorer(CFMENavigateStep):
+    VIEW = AutomateExplorerView
+    prerequisite = NavigateToSibling('LoggedIn')
+
+    def step(self):
+        self.view.navigation.select(*automate_menu_name(self.obj.appliance) + ['Explorer'])
 
 
 def check_tree_path(actual, desired):
